fix int

+++ marchalib/powspec.py
@@ -90,13 +90,13 @@ def powspec(image, reso=1, autocorr=False, **kwargs):
 	#Average values in same k bin
 	#---------------------------------------------
 
-	kval = np.zeros(np.int(k_crit))
-	kpow = np.zeros(np.int(k_crit))
+	kval = np.zeros(int(k_crit))
+	kpow = np.zeros(int(k_crit))
 
-	for j in range(np.int(k_crit)):
+	for j in range(int(k_crit)):
 
-		kval[j] = np.sum(k_mod[k_mod == np.float(j)]) / hval[j]
-		kpow[j] = np.sum(ps2D[k_mod == np.float(j)]) / hval[j]
+		kval[j] = np.sum(k_mod[k_mod == np.float64(j)]) / hval[j]
+		kpow[j] = np.sum(ps2D[k_mod == np.float64(j)]) / hval[j]
 
 	spec_k = kpow[1:np.size(hval)-1]
 
@@ -216,13 +216,13 @@ def cross_spec(image, image2, reso=1):
 	#Average values in same k bin
 	#---------------------------------------------
 
-	kval = np.zeros(np.int(k_crit))
-	kpow = np.zeros(np.int(k_crit))
+	kval = np.zeros(int(k_crit))
+	kpow = np.zeros(int(k_crit))
 
-	for j in range(np.int(k_crit)):
+	for j in range(int(k_crit)):
 
-		kval[j] = np.sum(k_mod[k_mod == np.float(j)]) / hval[j]
-		kpow[j] = np.sum(ps2D[k_mod == np.float(j)]) / hval[j]
+		kval[j] = np.sum(k_mod[k_mod == np.float64(j)]) / hval[j]
+		kpow[j] = np.sum(ps2D[k_mod == np.float64(j)]) / hval[j]
 
 	spec_k = kpow[1:np.size(hval)-1]
